added types of projectiles and optimized code

+++ source/actors/Actor.hx
@@ -1,6 +1,5 @@
 package actors;
 
-import flixel.FlxG;
 import flixel.FlxObject;
 import flixel.FlxSprite;
 import flixel.group.FlxGroup.FlxTypedGroup;
@@ -21,8 +20,6 @@ class Actor extends FlxSprite {
 	public var mA:Float;
 	public var indexOfselectedWeapon:Int;
 	public var isPlayer:Bool;
-	
-	public var bullets:FlxTypedGroup<Projectile>;
 
 	// Buffable properties
 	public static inline var SPEED = "speed";
@@ -43,7 +40,6 @@ class Actor extends FlxSprite {
 		passiveItems = new List<BaseItem>();
 
 		weapons = new FlxTypedGroup<BaseItem>(2);
-
 	}
 
 	override public function update(elapsed:Float):Void {
@@ -51,11 +47,13 @@ class Actor extends FlxSprite {
 	}
 
 	public static function takeDamage(p:Projectile, a:Actor):Void {
-		p.kill();
-		a.hurt(p.damage);
+		if ((p.owner.isPlayer && !a.isPlayer) || (!p.owner.isPlayer && a.isPlayer)) {
+			p.kill();
+			a.hurt(p.damage);
 
-		if (a.health <= 0) {
-			a.healthBar.kill();
+			if (a.health <= 0) {
+				a.healthBar.kill();
+			}
 		}
 	}
 }
